Provide Data on Potential Evaporation

Provide Data on Potential Evaporation

Usage

get_potential_evaporation(is_waterbody, district, lookup)

Arguments

is_waterbody

(vector of) logical indicating whether a block area is of type (from the land_type/veg_class/irrigation tuple) "waterbody"

district

(vector of) integer indicating the district number of the plot area (from the original input column "BEZIRK")

lookup

data frame with key columns is_waterbody, district and value columns etp, etps. A data frame of the required structure is returned by abimo_config_to_config in list element "potential_evaporation"

Examples

if (FALSE) { # \dontrun{
config <- abimo_config_to_config(kwb.abimo::read_config())
get_potential_evaporation(
  is_waterbody = TRUE,
  district = 1,
  lookup = config$potential_evaporation
)

get_potential_evaporation(
  is_waterbody = c(TRUE, TRUE, FALSE, FALSE),
  district = c(1, 99, 2, 99),
  lookup = config$potential_evaporation
)
} # }