Function to be called when using ABIMO with the 'old' Berlin-data structure

It calls the functions prepare_input_data and abimo_config_to_config and modifies the config-object to the required structure.

Usage

prepare_berlin_inputs(
  data_file,
  config_file = kwb.abimo::default_config(),
  data = NULL,
  config = NULL,
  dbg = TRUE
)

Arguments

data_file

a path for the .dbf file containing the input data

config_file

a path for the .xml config file

data

optional. Data frame containing the input data, as e.g. provided in abimo_input_2019.

config

optional. List representing an Abimo configuration, as e.g. provided by read_config.

dbg

logical indicating whether or not to show debug messages

Value

a list object containing elements data, the prepared input data and config, the prepared config object, both in a form readable by run_rabimo