R-Abimo Inputs (Data and Config) for Berlin, 2025.

Data frame and configuration object that are required by the R-Abimo main function run_rabimo. The data have been provided by Senate Department for Urban Development, Building and Housing III D Spatial Data Infrastructure, Environmental Atlas.

Usage

rabimo_inputs_2025

Format

rabimo_inputs_2025

A list containing two elements:

data

a data frame with the input data in R-Abimo format (see below)

config

a list object with configuration data (see below)

rabimo_inputs_2025$data

A data.frame with 58531 observations of 26 variables:

code

Unique block area identifier (character)

prec_yr

Long-term average of annual precipitation in mm (integer)

prec_s

Long-term average of annual precipitation within summer months (May to October) in mm (integer)

epot_yr

Long-term average of annual potential evapotranspiration in mm (integer)

epot_s

Long-term average of annual potential evapotranspiration within summer months (May to October) in mm (integer)

district

Number of Berlin "Bezirk" (district) in which the block area is located (character). This column is Berlin-specific and optional, i.e. not required by the model.

total_area

Total block area in square metres (numeric)

roof

Fraction of the total area that is considered as "roof" area (numeric value between 0.0 and 1.0)

green_roof

Fraction of the roof area that belongs to green roofs (numeric value between 0.0 and 1.0). A value of 1.0 means that all roofs in the block area are green roofs.

swg_roof

Fraction of the roof area that is connected to the drainage system (numeric value between 0.0 and 1.0)

pvd

Fraction of the total block area that is paved (numeric value between 0.0 and 1.0)

swg_pvd

Fraction of the paved area that is connected to the drainage system (numeric value between 0.0 and 1.0)

srf1_pvd

Fraction of the paved area that belongs to surface class 1 (numeric value between 0.0 and 1.0, see note 1 below)

srf2_pvd

Fraction of the paved area that belongs to surface class 2 (numeric value between 0.0 and 1.0, see note 1 below)

srf3_pvd

Fraction of the paved area that belongs to surface class 3 (numeric value between 0.0 and 1.0, see note 1 below)

srf4_pvd

Fraction of the paved area that belongs to surface class 4 (numeric value between 0.0 and 1.0, see note 1 below)

srf5_pvd

Fraction of the paved area that belongs to surface class 5 (numeric value between 0.0 and 1.0, see note 1 below)

to_swale

Fraction of sealed area (roof area + paved area) that is connected to an infiltration swale (numeric value between 0.0 and 1.0)

gw_dist

Distance between groundwater table and surface in metres (numeric)

ufc30

field capacity in 30 cm depth (numeric)

ufc150

field capacity in 150 cm depth (numeric)

land_type

land type, one of forested, horticultural, urban, vegetationless, waterbody (character)

veg_class

vegetation class index (numeric), derived from an analysis of tree volumes

irrigation

irrigation in mm per year (integer)

block_type

Block type identifier of the form "usage-type-id_block-type-id_usage-type-description_block-type-description" (character). This column is Berlin-specific and optional, i.e. not required by the model.

Shape

List structure containing geometry information on the different block areas. This column is optional. If provided, it will be appended to the model output so that model results can be plotted in the form of maps.

Note 1: The sum of surface class fractions srf1_pvd, srf2_pvd, srf3_pvd, srf4_pvd, srf5_pvd should be 1.0 within each block area.

rabimo_inputs_2025$config

A list with 3 named elements:

runoff_factors

Runoff factors for roofs and five different surface types, given as a vector of numeric with element names roof, surface1, surface2, surface3, surface4, surface5. A runoff factor determines the proportion of precipitation that, after subtraction of evapotranspiration, becomes surface runoff from a paved area. The higher the factor, the less permeable is the surface.

bagrov_values

Bagrov values to calculate evapotranspiration from paved surfaces, given as a vector of numeric with element names roof, green_roof, surface1, surface2, surface3, surface4, surface5. The higher the Bagrov value, the more evapotranspiration is generated by the model. For a description of the evapotranspiration model and for a figure that shows the influence of the Bagrov values (n) on the evapotranspiration (in German), see https://www.berlin.de/umweltatlas/wasser/wasserhaushalt/2001/methode/

swale

Model parameter(s) related to the 'swale' measure, given as a vector of numeric with currently one value, named swale_evaporation_factor. The swale evaporation factor determines which fraction of the water going into a swale becomes evapotranspiration (the rest becomes infiltration).

Source

https://www.berlin.de/umweltatlas/en/general/contact/

https://gdi.berlin.de/services/wfs/ua_gruendaecher_2020?REQUEST=GetCapabilities&SERVICE=wfs