Deviation from Natural Water Balance (Delta-W)

Calculate the deviation from the natural water balance (delta-W) given R-Abimo results as returned by run_rabimo.

Usage

calculate_delta_W(natural, urban, ..., implementation = 3L)

Arguments

natural

R-Abimo results for the "natural" scenario

urban

R-Abimo results for the "urban" scenario

...

further arguments passed to the implementation function as specified in implementation

implementation

one of 1, 2, 3, indicating a different implementation

Value

a data frame with the area codes in column code and the delta-W values in column delta_w