Define List of "Controls"

Define a list of settings that control how the main function run_rabimo should behave.

Usage

define_controls(
  check = TRUE,
  use_abimo_bagrov_solver = TRUE,
  reproduce_abimo_error = FALSE,
  output_format = "rabimo",
  intermediates = FALSE
)

Arguments

check

logical indicating whether the check functions are executed. Default: TRUE.

use_abimo_bagrov_solver

logical indicating whether or not to use the (fast!) algorithm implemented in Abimo to solve the Bagrov equations. Default: TRUE.

reproduce_abimo_error

logical indicating whether or not to reproduce the error that is contained in Abimo (missing area fraction factor). Default: FALSE.

output_format

one of "abimo" (upper case columns: CODE, R, ROW, RI, RVOL, ROWVOL, RIVOL, FLAECHE, VERDUNSTUN), "rabimo" (lower case columns: code, surface_runoff, infiltration, evaporation).

intermediates

logical indicating whether the intermediate results are returned as attributes. Default: FALSE.

Value

list with the arguments of this function as list elements

Examples

inputs <- kwb.rabimo::rabimo_inputs_2020
result <- run_rabimo(inputs$data, inputs$config)
#> Collecting climate related data ... ok. (0.00 secs) 
#> Preparing soil property data for all block areas ... ok. (1.16 secs) 
#> Precalculating actual evapotranspirations for impervious areas ... ok. (7.83 secs) 
#> Precalculating actual evapotranspirations for waterbodies or pervious areas ... ok. (0.63 secs) 
colMeans(as.matrix(result[, -1L]))
#>       area     runoff    infiltr     evapor 
#> 15225.0426   162.5073   184.4515   284.8178